Finalised adding purge,  react, and motto functionality
Separated functionality into respective js files, and added them to index so as to clean up code and allow for more modularity.
