+++ events/index.js
@@ -1,4 +1,7 @@
 module.exports = {
     "destroy" : require("./destroy.js"),
-    "who" : require("./who.js")
+    "who" : require("./who.js"),
+    "react" : require("./react.js"),
+    "purge" : require("./purge.js"),
+    "motto" : require("./motto.js")
 }
