+++ events/motto.js
@@ -1,0 +1,20 @@
+// Sleep function for async delay
+const sleep = (delay) => new Promise((resolve) => setTimeout(resolve, delay));
+
+// Sends a message, then edits itself to write out BBB
+async function execute(message) {
+    //Send "bella"
+    message.channel.send(':b:ella').then(async(sentMessage) => {
+        //Wait 1 second
+        await sleep(900);
+        //Edit the message to "Beans"
+        sentMessage.edit(":b:ella :b:eans").then(async(sentMessage) => {
+            //Wait 1 second
+            await sleep(900);
+            //Edit the message to "Bonstruction"
+            sentMessage.edit(":b:ella :b:eans :b:onstruction");
+        });
+    });
+}
+
+exports.execute = execute;
